backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
chromadb
sentence-transformers
pypdf
youtube-transcript-api
pytube
mysql-connector-python
python-dotenv